SWIPL2J.PrologStream

SWIPL2J.close

SWIPL2J.close

SWIPL2J.consult

SWIPL2J.create_query

SWIPL2J.echo_term

SWIPL2J.is_prolog_list

SWIPL2J.open_stream

SWIPL2J.parse_prolog_list_line

SWIPL2J.prompt_file_creation

SWIPL2J.query_all_values

SWIPL2J.query_bool

SWIPL2J.query_manual

SWIPL2J.query_swipl

SWIPL2J.query_value

SWIPL2J.read_end

SWIPL2J.read_prolog_error_alert

SWIPL2J.read_prolog_error_line

SWIPL2J.save

SWIPL2J.save_stream

SWIPL2J.start_swipl

SWIPL2J.start_swipl

SWIPL2J.try_create_file

SWIPL2J.unix_path

SWIPL2J.unload

SWIPL2J.write

SWIPL2J.write_swipl

SWIPL2J.PrologStream — Type

PrologStream

A stream object for a SWI-Prolog process.

Fields

alias::String: unique stream identifier

mode::Symbol: mode of stream (:read, :write, or :append)

filename::String: name of the active file for the stream

swipl::IO: reference to the swipl process in which the stream belongs to

SWIPL2J.close — Method

close(swipl)

Close the SWI-Prolog instance. https://www.swi-prolog.org/pldoc/doc_for?object=halt/0

Arguments

swipl: SWI-Prolog process

SWIPL2J.close — Method

close(stream::PrologStream)::Nothing

Closes a stream. streamproperty: https://www.swi-prolog.org/pldoc/docfor?object=streamproperty/2 close: https://www.swi-prolog.org/pldoc/docfor?object=close/1

Arguments

stream::PrologStream: stream object for an SWI-Prolog instance

SWIPL2J.consult — Function

consult(swipl, file::String, create_file::Bool = false)

Consult the file for SWI-prolog. https://www.swi-prolog.org/pldoc/doc_for?object=consult/1

Arguments

swipl: SWI-Prolog process

file::String: file to consult

create_file::Bool=false: silently create the file if it doesn't exist.

SWIPL2J.create_query — Method

create_query(query::String)::String

Turns the given string into a standardized query equipped to handle payload and final result.

https://www.swi-prolog.org/pldoc/docfor?object=writeln/1 https://www.swi-prolog.org/pldoc/docfor?object=flush_output/0

Arguments

query::String: a string which will be converted into a larger query

SWIPL2J.echo_term — Method

echo_term()

Open a terminal connection. https://docs.julialang.org/en/v1/base/io-network/

SWIPL2J.is_prolog_list — Method

is_prolog_list(string::String):Bool

returns whether a given string is a prolog list of the form list = [...].

Arguments

string::String: the string to be determined if it is a list

SWIPL2J.open_stream — Function

open_stream(swipl, file::String, mode::Symbol = :write, create_file::Bool = false, 
alias::Union{String, Nothing} = nothing)::Union{PrologStream, Nothing}

Opens a file via a stream for reading, writing, or appending. streamproperty: https://www.swi-prolog.org/pldoc/docfor?object=streamproperty/2 open: https://www.swi-prolog.org/pldoc/docfor?object=open/4

Arguments

swipl: SWI-Prolog process

file::String: active file for the stream

mode::Symbol=:write: mode in which the stream interacts with the file (:read, :write, or :append)

create_file::Bool=false: silently create the file if it doesn't exist.

alias::Union{String, Nothing}=nothing: unique string for SWI-Prolog to identify the stream, automatically

generated by default

SWIPL2J.parse_prolog_list_line — Method

parse_prolog_list_line(string::String)::Vector{String}

parses an SWI-Prolog list into a vecctor of strings.

Arguments

string::String: a list as a string

SWIPL2J.prompt_file_creation — Function

prompt_file_creation(file::String, create_file::Bool = false)::Bool

Helper loop to ask prompt the user to create a file.

Arguments

file::String: file to create

create_file::Bool=false: silently create the file if it doesn't exist

SWIPL2J.query_all_values — Method

query_all_values(swipl::IO, query::String)::Union{Vector{String}, Nothing}

queries SWI-Prolog and returns the results as a vector of strings. A query does not have to return multiple results, a non-list result will be wrapped as a string inside the vector. This function is suited for single line results

Arguments

swipl::IO: SWI-Prolog process

query::String: query to send to SWI-Prolog

SWIPL2J.query_bool — Method

query_bool(swipl::IO, query::String)::Union{Bool, Nothing}

Queries SWI-Prolog and returns the result of the query as a boolean, returns nothing if the query did not yeild true., or false..

Arguments

swipl::IO: SWI-Prolog process

query::String: query to send to SWI-Prolog

SWIPL2J.query_manual — Function

query_manual(swipl::IO, query::String, read_until_end=true)

Sends an unmodified query to SWI-Prolog, reads and returns however many lines passed in as the readlines parameter not including empty lines

Arguments

swipl::IO: SWI-Prolog process

query::String: query to send to prolog

read_until_end::Bool=true: number of lines to read after querying, ensure this number is not

greater than the number of expected lines out the query should return

SWIPL2J.query_swipl — Method

query_swipl(swipl::IO, query::String, debug::Bool=false)::NamedTuple{payload::Vector{String}, result::SubString{String}}

Sends a query to an SWI-Prolog process, returns a payload and result, if an error occurrs as a result of the query, the payload will house these error messages.

Arguments

swipl::IO: SWI-Prolog process

query::String: query to send to SWI-Prolog

debug::Bool=false: if true, prints payload.

SWIPL2J.query_value — Method

query_value(swipl::IO, query::String)::Union{String, Nothing}

Queries SWI-Prolog and returns the result of the query as a string, returns nothing if an error occured and the query yielded no result.

Arguments

swipl::IO: SWI-Prolog process

query::String: query to send to SWI-Prolog

SWIPL2J.read_end — Method

read_end(swipl::IO)::Nothing

Reads the lines containing the ending message, the ending message this line reads is how the program confirms a queries output has ended

Arguments:

-swipl::IO: an SWI-Prolog process

SWIPL2J.read_prolog_error_alert — Function

read_prolog_error_alert(query::Union{String, Nothing} = nothing)

prints a line alerting the user of an SWI-Prolog error, optionally prints the query causing the error.

#Arguments

-query::Union{String, Nothing}=nothing: the query which caused the error, if left as nothing, the alert will not specify any query

SWIPL2J.read_prolog_error_line — Method

read_prolog_error_line(query::Union{String, Nothing} = nothing)

prints a line with an SWI-Prolog error.

#Arguments

-string::String: a line, typially involved in the query

SWIPL2J.save — Function

save(swipl, file::String, create_file::Bool = false)

Saves the current facts from SWI-Prologs memory into the given file. tell: https://www.swi-prolog.org/pldoc/docfor?object=tell/1 listing: https://www.swi-prolog.org/pldoc/docfor?object=listing/0 told: https://www.swi-prolog.org/pldoc/doc_for?object=told/0

Arguments

swipl: SWI-Prolog process

file::String: file to save SWI-Prolog's current memory to

create_file::Bool=false: silently create the file if it doesn't exist.

SWIPL2J.save_stream — Method

save(stream::PrologStream)

Save any writes in memory to the streams file. https://www.swi-prolog.org/pldoc/docfor?object=flushoutput/1

Arguments

stream::PrologStream: stream object for an SWI-Prolog instance

SWIPL2J.start_swipl — Function

start_swipl(file, create_file)::Union{IO, Nothing}

Open and return an SWI-Prolog instance with an open file.

Arguments

file::String: name of the file to open the SWI-Prolog context with.

create_file::Bool=false: silently create the file if it doesn't exist.

SWIPL2J.start_swipl — Method

start_swipl()

Open and returns an SWI-Prolog process.

SWIPL2J.try_create_file — Method

try_create_file(file::String, dir::String)::Bool

Attempts to create a file at the given directory, returns a boolean status of the file creation.

Arguments

file::String: file to create

SWIPL2J.unix_path — Method

unix_path(path::String)::String

Replaces \ with / to prevent pathing errors

Arguments

path::String: file path

SWIPL2J.unload — Method

unload(swipl, file::String)

Unload the file as to not include its context for future queries. https://www.swi-prolog.org/pldoc/docfor?object=unloadfile/1

Arguments

swipl: SWI-Prolog process

file::String: file to unload

SWIPL2J.write — Method

write(stream::PrologStream, message::String)

Write into a stream. The result of write depends on the mode of the stream, :append will add the text to the end of the file, :write will replace the existing file contents with the message. Writing to a stream writes directly into the file, thus using the save function is not required. https://www.swi-prolog.org/pldoc/doc_for?object=write/2

Arguments

stream::PrologStream: stream object for an SWI-Prolog instance

message::String: text to write into the stream

SWIPL2J.write_swipl — Method

write_swipl(swipl, query::String)::Nothing

(Incomplete Function) Write a line to an SWI-Prolog process.

Arguments

swipl: SWI-Prolog process

query::String: message to be written to the SWI-Prolog process.